examples/research/requirements.txt
langgraph-cli[inmem]
tavily-python
